Updating testing login process 2
Ok when logging into the database the ConnectionManager class needs a
user name and password to login to MySql. Now the user name and password
are Final Static Class members of the class FinalStaticLoginInfo . This
class is in the .gitignore file now so you can change it to your user
name and password.